% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatty.R
\name{chatty}
\alias{chatty}
\title{Make a variable chatty}
\usage{
chatty(
  ...,
  f = identity,
  silent.access = FALSE,
  silent.modification = FALSE,
  deep = FALSE
)
}
\arguments{
\item{...}{variables to make chatty}

\item{f}{function to used on modified values before printing them, by default
the full modified object is printed}

\item{silent.access}{Whether to make access silent}

\item{silent.modification}{Whether to make modifications silent}

\item{deep}{Whether to make function arguments chatty when function are passed
a chatty variable as an input}
}
\value{
Returns \code{NULL} invisibly, called for side effects.
}
\description{
Call \code{chatty(x, y)} to make the variables \code{x} and \code{y} chatty. A message
will then be printed every time they are accessed or modified. If you set
deep to true, when the variable is given directly as an argument to a function
(as in \code{fun(x)} but not \code{fun(x + 1)} the argument will become chatty again.
}
