% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_ambiguous_summarize_calls.R
\name{detect_ambiguous_summarize_calls}
\alias{detect_ambiguous_summarize_calls}
\title{Detect ambiguous \code{summarize()} calls}
\usage{
detect_ambiguous_summarize_calls(paths = ".", recursive = TRUE, pattern = NULL)
}
\arguments{
\item{paths}{Paths to scripts or folders containing scripts}

\item{recursive}{Whether folders should be explorer recursively}

\item{pattern}{A regular expression used to filter files}
}
\description{
Detect ambiguous \code{summarize()} calls
}
